skapade en liknande metod för att redigera en post som för radering av post.
all info om posten skickas med med metoden(namn, desc osv. )

+++ NoteItem.cs
@@ -70,13 +70,42 @@ namespace todo
 
         private void buttonDelete_Click(object sender, EventArgs e)
         {
-            //skriver ut id för note i console
-            Console.WriteLine("button " + this.Id + " clicked");
+            //en extra ruta som frågar om man verkligen vill radera note
+            DialogResult dialogResult = MessageBox.Show("Radera vald note?", "Radera", MessageBoxButtons.YesNo);
+            if (dialogResult == DialogResult.Yes)
+            {
+                //skriver ut id för note i console
+                Console.WriteLine("button " + this.Id + " clicked");
 
-            int id= Int32.Parse(this.Id);
+                //tilldelar rätt id och konverterar till int
+                int id= Int32.Parse(this.Id);
+                Form1 form = new Form1();
+
+                //kallar funktion som finns på main form och skickar med id så att rätt post raderas
+                //funktionen tar bort note från tabellen
+                form.DeleteNote(id);
+
+                //tar bort controllern från flowlayoutpanelen
+                this.Parent.Controls.Remove(this);
+            }
+            else if (dialogResult == DialogResult.No)
+            {
+
+            }
+        }
+
+        private void buttonEdit_Click(object sender, EventArgs e)
+        {
+            //tilldelar rätt id och konverterar till int
+            int id = Int32.Parse(this.Id);
+            string name = this.Name;
+            string desc = this.Description;
+            string color = this.ColorCode;
 
             Form1 form = new Form1();
-            form.DeleteNote(id);
+            form.editNote(id, name, desc, color);
+
+
         }
     }
 }
